Add a js file
I add a js file contain four function  that will trigger the timer interval when user is clicked on the play button. The time will be decreased until it equal 0, and display a message that tell user to do something. There is a reload button when user want to start over the count down time. There are more things that I will need to add to the html file.